+++ src/main/java/com/napier/sem/CitiesPopulationReport.java
@@ -6,92 +6,16 @@ import java.sql.ResultSet;
 import java.util.ArrayList;
 import java.util.List;
 
-public class CityPopulationReport {
+public class CitiesPopulationReport {
 
     private final Connection con;
 
-    public CityPopulationReport(Connection con) {
+    public CitiesPopulationReport(Connection con) {
         this.con = con;
     }
 
     Helpers helpers = new Helpers();
 
-    public void generateTopNPopulatedCapitalCitiesInTheWorldReport(int N) {
-
-        // SQL query to select the top N populated capital cities in the world
-        String query =
-                "SELECT city.Name, country.Name AS Country, city.Population " +
-                        "FROM city JOIN country ON city.ID = country.Capital " +
-                        "ORDER BY city.Population DESC LIMIT "+ N ;
-        ArrayList<City> cities = generateCityData(query);
-
-        // Prepare data for printing
-        String title = "Top " + N + " Populated Capital Cities in the World";
-        List<String> columnNames = List.of("City Name", "Country", "Population");
-        List<List<String>> rows = new ArrayList<>();
-
-        for (City city: cities) {
-            List<String> row = new ArrayList<>();
-            row.add(city.getName());
-            row.add(city.getCountryCode());
-            row.add(String.valueOf(city.getPopulation()));
-            rows.add(row);
-        }
-
-        helpers.printReport(title, columnNames, rows);
-    }
-
-    public void generateTopNPopulatedCapitalCitiesInTheRegionReport(int N, String region) {
-
-        // SQL query to select the top N populated capital cities in the region
-        String query =
-                "SELECT city.Name, country.Name AS Country, city.Population " +
-                        "FROM city JOIN country ON city.ID = country.Capital " +
-                        "WHERE country.Region='" + region + "' " +
-                        "ORDER BY city.Population DESC LIMIT "+ N ;
-        ArrayList<City> cities = generateCityData(query);
-
-        // Prepare data for printing
-        String title = "Top " + N + " Populated Capital Cities in region " + region;
-        List<String> columnNames = List.of("City Name", "Country", "Population");
-        List<List<String>> rows = new ArrayList<>();
-
-        for (City city: cities) {
-            List<String> row = new ArrayList<>();
-            row.add(city.getName());
-            row.add(city.getCountryCode());
-            row.add(String.valueOf(city.getPopulation()));
-            rows.add(row);
-        }
-
-        helpers.printReport(title, columnNames, rows);
-    }
-
-    public void generateTopNPopulatedCapitalCitiesInTheContinentReport(int N, String continent) {
-
-        // SQL query to select the top N populated capital cities in the continent
-        String query =
-                "SELECT city.Name, country.Name AS Country, city.Population " +
-                        "FROM city JOIN country ON city.ID = country.Capital " +
-                        "WHERE country.Continent='" + continent + "' " +
-                        "ORDER BY city.Population DESC LIMIT "+ N ;
-        ArrayList<City> cities = generateCityData(query);
-
-        // Prepare data for printing
-        String title = "Top " + N + " Populated Capital Cities in continent " + continent;
-        List<String> columnNames = List.of("City Name", "Country", "Population");
-        List<List<String>> rows = new ArrayList<>();
-
-        for (City city: cities) {
-            List<String> row = new ArrayList<>();
-            row.add(city.getName());
-            row.add(city.getCountryCode());
-            row.add(String.valueOf(city.getPopulation()));
-            rows.add(row);
-        }
-
-        helpers.printReport(title, columnNames, rows);
-    }
 
     public void generatePopulationInCitiesVSNonCityByContinent() {
         // SQL query to compute population data by continent
