+++ src/test/java/com/napier/sem/CitiesPopulationReportTest.java
@@ -21,7 +21,7 @@ import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)
-public class CityPopulationReportTest {
+public class CitiesPopulationReportTest {
 
     @Mock
     private Connection con;
@@ -33,7 +33,7 @@ public class CityPopulationReportTest {
     private ResultSet rset;
 
     @InjectMocks
-    private CityPopulationReport cityPopulationReport;
+    private CitiesPopulationReport cityPopulationReport;
 
     private final ByteArrayOutputStream outContent = new ByteArrayOutputStream();
 
@@ -45,90 +45,6 @@ public class CityPopulationReportTest {
         System.setOut(new PrintStream(outContent));
         when(con.prepareStatement(anyString())).thenReturn(stmt);
         when(stmt.executeQuery()).thenReturn(rset);
-    }
-
-    @Test
-    void testGenerateTopNPopulatedCapitalCitiesInTheWorldReport() throws Exception {
-        when(rset.next()).thenReturn(true, true, true, true, true, false); // Simulate five rows returned
-        when(rset.getString(eq("Name"))).thenReturn("Beijing");
-        when(rset.getString(eq("Country"))).thenReturn("China");
-        when(rset.getInt("Population")).thenReturn(100000);
-
-        int limit = 5;
-        cityPopulationReport.generateTopNPopulatedCapitalCitiesInTheWorldReport(limit);
-
-        verify(con).prepareStatement(sqlCaptor.capture());
-        String executedSQL = sqlCaptor.getValue();
-        String expectedQuery = "SELECT city.Name, country.Name AS Country, city.Population " +
-                        "FROM city JOIN country ON city.ID = country.Capital " +
-                        "ORDER BY city.Population DESC LIMIT "+ limit ;
-
-        assertTrue(executedSQL.contains(expectedQuery),
-                "The SQL should contain the correct LIMIT clause for the capital city.");
-
-        // Verify printReport invocation indirectly by checking the console output
-        String output = outContent.toString();
-        assertTrue(output.contains("Top " + limit + " Populated Capital Cities in the World"), "Output should contain the report title.");
-        assertTrue(output.replaceAll("\\s\\s+", "\t").contains("City Name\tCountry\tPopulation"),
-                "Output should contain the correct column headers.");
-    }
-
-    @Test
-    void testGenerateTopNPopulatedCapitalCitiesInTheRegionReport() throws Exception {
-        when(rset.next()).thenReturn(true, true, true, true, true, false); // Simulate five rows returned
-        when(rset.getString(eq("Name"))).thenReturn("Washington DC");
-        when(rset.getString(eq("Country"))).thenReturn("USA");
-        when(rset.getInt("Population")).thenReturn(100000);
-
-        String region = "North America";
-        int limit = 5;
-        cityPopulationReport.generateTopNPopulatedCapitalCitiesInTheRegionReport(limit, region);
-
-        verify(con).prepareStatement(sqlCaptor.capture());
-        String executedSQL = sqlCaptor.getValue();
-        String expectedQuery = "SELECT city.Name, country.Name AS Country, city.Population " +
-                "FROM city JOIN country ON city.ID = country.Capital " +
-                "WHERE country.Region='" + region + "' " +
-                "ORDER BY city.Population DESC LIMIT "+ limit ;
-
-        assertTrue(executedSQL.contains(expectedQuery),
-                "The SQL should contain the correct LIMIT clause for the capital city.");
-
-        // Verify printReport invocation indirectly by checking the console output
-        String output = outContent.toString();
-        assertTrue(output.contains("Top " + limit + " Populated Capital Cities in region " + region),
-                "Output should contain the report title.");
-        assertTrue(output.replaceAll("\\s\\s+", "\t").contains("City Name\tCountry\tPopulation"),
-                "Output should contain the correct column headers.");
-    }
-
-    @Test
-    void testGenerateTopNPopulatedCapitalCitiesInTheContinentReport() throws Exception {
-        when(rset.next()).thenReturn(true, true, true, true, true, false); // Simulate five rows returned
-        when(rset.getString(eq("Name"))).thenReturn("Beijing");
-        when(rset.getString(eq("Country"))).thenReturn("China");
-        when(rset.getInt("Population")).thenReturn(100000);
-
-        String continent = "Asia";
-        int limit = 5;
-        cityPopulationReport.generateTopNPopulatedCapitalCitiesInTheContinentReport(limit, continent);
-
-        verify(con).prepareStatement(sqlCaptor.capture());
-        String executedSQL = sqlCaptor.getValue();
-        String expectedQuery = "SELECT city.Name, country.Name AS Country, city.Population " +
-                "FROM city JOIN country ON city.ID = country.Capital " +
-                "WHERE country.Continent='" + continent + "' " +
-                "ORDER BY city.Population DESC LIMIT "+ limit ;
-
-        assertTrue(executedSQL.contains(expectedQuery),
-                "The SQL should contain the correct LIMIT clause for the capital city.");
-
-        // Verify printReport invocation indirectly by checking the console output
-        String output = outContent.toString();
-        assertTrue(output.contains("Top " + limit + " Populated Capital Cities in continent " + continent),
-                "Output should contain the report title.");
-        assertTrue(output.replaceAll("\\s\\s+", "\t").contains("City Name\tCountry\tPopulation"),
-                "Output should contain the correct column headers.");
     }
 
     @Test
